+++ Runtime/Authoring/AnimationStateMachine/AnimationParameterAsset.cs
@@ -1,11 +1,10 @@
 using Unity.Entities;
 using UnityEngine;
 
-namespace DOTSAnimation.Authoring
+namespace DMotion.Authoring
 {
-    public abstract class AnimationParameterAsset : ScriptableObject
+    public abstract class AnimationParameterAsset : StateMachineSubAsset
     {
-        public string Name;
-        public int Hash => Name.GetHashCode();
+        public int Hash => name.GetHashCode();
     }
 }