+++ Runtime/Authoring/AnimationStateMachine/FloatParameterAsset.cs
@@ -1,8 +1,7 @@
 using UnityEngine;
 
-namespace DOTSAnimation.Authoring
+namespace DMotion.Authoring
 {
-    [CreateAssetMenu(menuName = "DOTSAnimation/Parameters/Float Parameter")]
     public class FloatParameterAsset : AnimationParameterAsset
     {
     }