+++ Runtime/Authoring/AnimationStateMachine/LinearBlendStateAsset.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
 
-namespace DOTSAnimation.Authoring
+namespace DMotion.Authoring
 {
     [Serializable]
     public struct ClipWithThreshold
@@ -12,10 +12,9 @@ namespace DOTSAnimation.Authoring
         public float Threshold;
     }
     
-    [CreateAssetMenu(menuName = "DOTSAnimation/States/Blend Tree 1D")]
     public class LinearBlendStateAsset : AnimationStateAsset
     {
-        public ClipWithThreshold[] BlendClips;
+        public ClipWithThreshold[] BlendClips = Array.Empty<ClipWithThreshold>();
         public FloatParameterAsset BlendParameter;
         
         public override StateType Type => StateType.LinearBlend;