+++ Runtime/Authoring/AnimationStateMachine/SingleClipStateAsset.cs
@@ -1,9 +1,8 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-namespace DOTSAnimation.Authoring
+namespace DMotion.Authoring
 {
-    [CreateAssetMenu(menuName = "DOTSAnimation/States/Single Clip")]
     public class SingleClipStateAsset : AnimationStateAsset
     {
         public AnimationClipAsset Clip;