+++ Runtime/Authoring/AnimationStateMachine/StateMachineAsset.cs
@@ -2,26 +2,17 @@ using System.Collections.Generic;
 using System.Linq;
 using UnityEngine;
 
-namespace DOTSAnimation.Authoring
+namespace DMotion.Authoring
 {
-    [CreateAssetMenu(menuName = "DOTSAnimation/State Machine")]
+    [CreateAssetMenu(menuName = StateMachineEditorConstants.DMotionPath + "/State Machine")]
     public class StateMachineAsset : ScriptableObject
     {
         public AnimationStateAsset DefaultState;
-        public List<SingleClipStateAsset> SingleClipStates;
-        public List<LinearBlendStateAsset> LinearBlendStates;
-        public List<BoolParameterAsset> BoolParameters;
-        public List<FloatParameterAsset> FloatParameters;
-        public List<AnimationTransitionGroup> Transitions;
-        public IEnumerable<AnimationClipAsset> Clips => SingleClipStates
-            .SelectMany(s => s.Clips)
-            .Concat(LinearBlendStates.SelectMany(s => s.Clips));
-        
-        public int ClipCount => SingleClipStates.Sum(s => s.ClipCount) +
-                                LinearBlendStates.Sum(s => s.ClipCount);
-        public int StateCount => SingleClipStates.Count + LinearBlendStates.Count;
+        public List<AnimationStateAsset> States = new List<AnimationStateAsset>();
+        public List<AnimationParameterAsset> Parameters = new List<AnimationParameterAsset>();
+        public List<AnimationClipAsset> Clips = new List<AnimationClipAsset>();
+        public int ClipCount => Clips.Count;
 
-        public IEnumerable<AnimationStateAsset> States => SingleClipStates
-            .Concat(LinearBlendStates.OfType<AnimationStateAsset>());
+        public GameObject ClipPreviewGameObject;
     }
 }