+++ Runtime/Components/AnimationEvent.cs
@@ -1,11 +1,14 @@
 ﻿using System;
+using Latios.Kinemation;
 using Unity.Entities;
 
-namespace DOTSAnimation
+namespace DMotion
 {
     public struct RaisedAnimationEvent : IBufferElementData
     {
         public int EventHash;
+        public float ClipWeight;
+        public SkeletonClipHandle ClipHandle;
     }
     
     internal struct AnimationClipEvent