+++ Runtime/Components/AnimationStateMachine.cs
@@ -1,7 +1,7 @@
 using Latios.Kinemation;
 using Unity.Entities;
 
-namespace DOTSAnimation
+namespace DMotion
 {
     internal struct AnimationStateMachine : IComponentData
     {
@@ -10,11 +10,7 @@ namespace DOTSAnimation
         internal BlobAssetReference<StateMachineBlob> StateMachineBlob;
         internal AnimationState CurrentState;
         internal AnimationState NextState;
-        internal StateTransition CurrentTransition;
         internal float Weight;
-        
-        //TODO (perf): Do those get inlined? It's just syntax sugar
-        internal readonly ref AnimationTransitionGroup CurrentTransitionBlob =>
-            ref StateMachineBlob.Value.Transitions[CurrentTransition.TransitionIndex];
+        internal float CurrentTransitionNormalizedTime;
     }
 }