+++ Samples~/All Samples/1 - Complete State Machine/Scripts/UI/StateMachineExampleUI.cs
@@ -1,0 +1,15 @@
+﻿using TMPro;
+using Unity.Entities;
+using UnityEngine.UI;
+
+namespace DMotion.Samples.CompleteStateMachine
+{
+    [GenerateAuthoringComponent]
+    public class StateMachineExampleUI : IComponentData
+    {
+        public Slider BlendSlider;
+        public Button AtkButton;
+        public Button AtkRmButton;
+        public Toggle IsFallingToggle;
+    }
+}